Tests currency exchange service.

+++ net/Poe2scout/Repositories/PriceLog/Models/ItemPriceInRange.cs
@@ -2,5 +2,5 @@ namespace Poe2scout.Repositories.PriceLog.Models;
 
 public record ItemPriceInRange(
   int ItemId,
-  decimal Price,
-  decimal Quantity);
+  double Price,
+  int Quantity);
